requirements.txt
PyYAML>=6.0
requests>=2.32.0
Jinja2>=3.1.0
openai>=1.35.0
Flask>=3.0.0
